+++ deeplink/src/main/java/com/lyft/deeplink/DeepLinkParams.java
@@ -6,7 +6,7 @@ public class DeepLinkParams {
 
     private final String clientId;
     private final String rideType;
-    private final String couponCode;
+    private final String promoCode;
     private final String pickupAddr;
     private final Double pickupLat;
     private final Double pickupLng;
@@ -14,11 +14,11 @@ public class DeepLinkParams {
     private final Double dropoffLat;
     private final Double dropoffLng;
 
-    private DeepLinkParams(String clientId, String rideType, String couponCode, String pickupAddr, Double pickupLat, Double pickupLng,
+    private DeepLinkParams(String clientId, String rideType, String promoCode, String pickupAddr, Double pickupLat, Double pickupLng,
             String dropoffAddr, Double dropoffLat, Double dropoffLng) {
         this.clientId = clientId;
         this.rideType = rideType;
-        this.couponCode = couponCode;
+        this.promoCode = promoCode;
         this.pickupAddr = pickupAddr;
         this.pickupLat = pickupLat;
         this.pickupLng = pickupLng;
@@ -38,8 +38,8 @@ public class DeepLinkParams {
     }
 
     @Nullable
-    public String getCouponCode() {
-        return couponCode;
+    public String getPromoCode() {
+        return promoCode;
     }
 
     @Nullable
@@ -92,7 +92,7 @@ public class DeepLinkParams {
 
         private String clientId;
         private String rideType = "lyft";
-        private String couponCode;
+        private String promoCode;
         private String pickupAddr;
         private Double pickupLat;
         private Double pickupLng;
@@ -101,7 +101,7 @@ public class DeepLinkParams {
         private Double dropoffLng;
 
         public DeepLinkParams build() {
-            return new DeepLinkParams(clientId, rideType, couponCode, pickupAddr, pickupLat, pickupLng, dropoffAddr, dropoffLat, dropoffLng);
+            return new DeepLinkParams(clientId, rideType, promoCode, pickupAddr, pickupLat, pickupLng, dropoffAddr, dropoffLat, dropoffLng);
         }
 
         public Builder setClientId(String clientId) {
@@ -114,8 +114,8 @@ public class DeepLinkParams {
             return this;
         }
 
-        public Builder setCouponCode(String couponCode) {
-            this.couponCode = couponCode;
+        public Builder setPromoCode(String promoCode) {
+            this.promoCode = promoCode;
             return this;
         }