+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+Version 1.0.3 *(2017-02-21)*
+----------------------------
+Rename couponCode to promoCode
+Add promoCode to RideParams
+Make ResultLoadedCallback interface public
+
 Version 1.0.2 *(2016-11-22)*
 ----------------------------
 Coupon code parameter in ridetype deeplink